Update of jquery.pin.js
I've added a statement to check whether the scrolling content is taller than the static content. If it is don't scroll the content. I ran into this issue when implementing the jquery.pin.js plugin into a CMS and as content can vary on occasion the scrollable content was actually taller (height) than the static content. This caused the scroll content to constantly change from fixed to absolute and appeared to flicker.